aadhar/requirements.txt
fastapi
uvicorn
google-generativeai
PyMuPDF
Pillow
